+++ http_server/backend_epoll/tls_conn_linux.c.v
@@ -48,9 +48,9 @@ fn tls_set_out(mut conn TlsConn, epoll_fd int, fd int, want_out bool) {
 	}
 	conn.watching_out = want_out
 	if want_out {
-		epoll.mod_fd_in_epoll(epoll_fd, fd, u32(C.EPOLLIN | C.EPOLLOUT | C.EPOLLET))
+		epoll.mod_fd_in_epoll(epoll_fd, fd, (u32(C.EPOLLIN) | u32(C.EPOLLOUT) | u32(C.EPOLLET)))
 	} else {
-		epoll.mod_fd_in_epoll(epoll_fd, fd, u32(C.EPOLLIN | C.EPOLLET))
+		epoll.mod_fd_in_epoll(epoll_fd, fd, (u32(C.EPOLLIN) | u32(C.EPOLLET)))
 	}
 }
